# Improve watch performance by using stable identities for Snapshot iterables