@JoostK 2022-06-30
I believe I saw the same bottleneck before, then reported as https://github.com/webpack/webpack/issues/15231.