@gluxon 2022-06-30
@JoostK Thanks for linking that. I was curious if others had seen this issue. I would be interested to know if the patch here alleviates the problem for your team if you have a moment to manually apply and test.

While this PR brought our team's `merge` time down from 1 minute, I did notice times still vary from 1s to 4s. I think `Snapshot` can non-deterministically end up in a deeply nested state that takes `SnapshotIterator` a bit of time to flatten. I was hoping to continue investigating there, but I'm not sure if it's worthwhile yet unless it would help others.