@JoostK 2022-06-30
I just checked with this patch applied and it didn't _seem_ any faster. I haven't taken measurements but it doesn't seem like it resolves #15231.