@gluxon 2022-07-01
Got it. I'd wager there's more work needed here too. The next step I was thinking of was attempting to serialize the `Snapshot` tree and comparing it between the fast ~1s `merge` case and slow ~4s case.

As a heads up, the `watch` -> `merge` slowness this PR addresses can be difficult to perceive without measuring/profiling. The watch call happens after Webpack says it's ready/done, so the timings here aren't typically printed on the console. If you have time, I would be curious if the `lib/Watching.js` timings are different in @mattlewis92's profiling here: https://github.com/webpack/webpack/issues/15231#issue-1111649057