@sokra 2022-07-01
> * Does this change make sense, or should I investigate improving performance of this `watch` call in a different way?

Makes sense to me

> * Would it make sense to use [`WeakRef`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WeakRef) to prevent these saved iterable objects from living longer than necessary? I think we'd have to check for the API's existence and gracefully fallback to support older Node.js versions.

I don't think that makes sense, since each iterator only keeps a referenced to the Snapshot, so it doesn't add any large memory overhead. Actually it reduces memory since only a single iterator is created per Snapshot.

> Items iterated during merge: 1198677, 23045.103875994682 ms

> Items iterated during merge: 464918, 1726.2539209723473 ms

It's interesting that reducing the count by 50% improves performance by 13x. Why is that? Maybe the SnapshotIterables are far more expensive than the remaining Iterables during merge, and the number of SnapshotIterables was also reduced by ~13x.

> I think `Snapshot` can non-deterministically end up in a deeply nested state

The way `Snapshot`s are optimized is really non-deterministic, since it depends a lot on the order and speed in creating them (which depends on fs operations).

> I contacted my company's CLA management team to work out the EasyCLA.

:+1:

The PR looks good and I'll merge it afterwards