@gluxon 2022-07-01
> I don't think that makes sense, since each iterator only keeps a referenced to the Snapshot, so it doesn't add any large memory overhead. Actually it reduces memory since only a single iterator is created per Snapshot.

Nice. Glad to have a second pair of eyes to confirm that.

> It's interesting that reducing the count by 50% improves performance by 13x. Why is that? Maybe the SnapshotIterables are far more expensive than the remaining Iterables during merge, and the number of SnapshotIterables was also reduced by ~13x.

I was surprised by this too and came to the same conclusion as you. I think the `SnapshotIterable` objects that were duplicated were typically more nested/larger than the more unique objects.

I hadn't thought of this until your memory comment just now, but I'd expect garbage collection of `SnapshotIterable` objects to happen less in this `watch` -> `merge` call now. There may have been performance improvements from just not running GC as frequently.

> The way Snapshots are optimized is really non-deterministic, since it depends a lot on the order and speed in creating them (which depends on fs operations).

👍 Thanks for confirming that. That's nice to know.