@gluxon 2022-07-01
Are the codecov status errors something that's been seen before in the Webpack repo?

<img width="796" alt="Screen Shot 2022-07-01 at 3 56 24 PM" src="https://user-images.githubusercontent.com/906558/176961032-1204a1d7-e2cd-43f8-a75f-0070e2d9536f.png">

It is possible that the changes here affected code paths in tests, but wanted to check if this was seen before investigating how my changes would have affected logic in `lib/Compiler.js` and `lib/Compilation.js`.